move schema.ts to a separate schema/ dir and move the present notes to a readme
